[#11138] refactor(iceberg-rest): move federation table and credential handling into FederatedCatalogWrapper
